Bound hibernate.hbm2ddl.auto to a maven property defaulted to previous fixed value.
Parameterized "hibernate.hbm2ddl.auto" property, with the pom.xml value
matching the original fixed value of "update".  When running lots of
Hibernate-dependent test runs to explore code changes that are not
schema-modifying, a developer may wish to tune their Maven profile to
set hibernate.hbm2ddl.auto == 'none', so Hibernate doesn't spend so much
time verifying the schema over and over and over and over and over.

For that matter, even with changes that are schema-mutating, there may
be good reason to disable the schema update checks once those have
occurred and work has progressed to dealing with the ramifications on
Repository layer classes and higher up the stack.