Made mobsters-cache executable
Made a few tweaks to make it possible to execute some ground-work for
reading the 20K or so of configuration data into memory and retrieving
it by indexed keys.
-- Fixed a typo in the spring-jedis.xml file (factoryMethod instead of
factory_method)
-- Added a missing dependency on commons-pool that somehow is not
acquired transitively from Jedis and/or Krypto.
-- Transformed BasePersistentObject to expect Integer keys instead of
Strings since MySQL is not expected to transform its existing integer ID
keys to Strings like DynamoDB-destined tables will.
-- Modified a handful of constructors to ensure the int id argument
given to the full-form constructor actually gets used

This also includes some experimental changes made to Clan to assert it
is possible to leverage Hibernate functionality to navigate Id-links in
MySQL config tables without maintaining ID-based lookup tables and A->B
utlity mapping methods.  It seems to work as it always has, and presents
an opportunity for a code-reducing shortcut.  The trick was getting
ForeignKey dependence to stop being a requirement.