Propertified all versions in root pom.
Propertified all version dependencies in root pom.xml.  This makes it a
little easier for a developer to experiment with an updated version
without having to hack the pom instead of their local mvnSettings.xml.

Still TBD: Maven plugins should acquire explicit ${maven-XXX.plugin-
version} properties in version declarations so builds are deterministic
instead of always relying on the latest version of each plugin working
the same as the last any time a newer version is released.