Removing unneeded and illegal class hierarchy from JPA Entity set
PrivateChatPost/AdminChatPost create an illegal Entity sub-tree as-is.
They aren't likely to remain in MySQL, so rather than attempt to fix the
ORM, its easier to just disable it.  Entity classes and repository
interfaces will likely be removed altoghether later on unless this turns
out to be an unexpectedly bad assumption.