CrudRepository->JpaRepository
Modified several repositories from CrudRepository inheritance to
JpaRepository inheritance in order to gain use of the ***InBatch()
methods that process collections as a single multi-key query rather than
N single-key statements as well as the modified save() semantics that
defers query execution until transaction-end or explicit flush(), for
the findAll() variants that return a List rather than an Iterable.

Perhaps most importantly of all, the JpaRepository interface can handle
@EntityGraph annotations that allow the default fetch types to be
overriden and the scope of eager fetching to be intentionally and
selectively broadened.

Second-to-most-importantly, the getById() methods that enable a proxy
for an non-fetched object to be retrieved for use in establishing
references to it by its Id.  The object is not fetched uless its lazy
initialization is triggered by reading a non-Id property.  The validity
of the Id link is performed server-side, within the RDBMS when the
attempt to save a linking object is flush()ed.