Added a project for reusable data harnesses
For integration level tests, a facility for reusing subunits of
bootstrap data and re-setting the envirionment to a known state is
helpful.  This is a rudimentary first such project which carries a
dependency on mobsters-sql, which it presumes to be in a sane and
"tested" state, and provided services for bootstrapping named building
blocks of objects with known configuration.

As of yet, the modularity is there but the facilities for selectively
turning a subunit on or off is not.  The intent is to tie this to Spring
profiles so that a test fixture can select blocks of data dependencies
as:

@ActiveProfiles({"SampleUserData", "SampleMonsterData",
"MonsterInstancesForSamAndJim"})

Where, in this case, the SampleUserData subunit defines two users, Sam
and Jim, and MonsterInstancesForSamAndJim instantiates a handful of
monsters from SampleMonsterDAta for them.

TBD is handling for dependencies, such that including the latter of the
three examples above necessarily includes the former two and is
initialized after them.

This may or may not be useful, but for a first few test cases it's DRY
for now.