Added reflective BaseDynamoRepo.createTable()
Rather than declaring a createTable method that repeats and must be kept
consistent with the DynamoDBMapper annotation set, this replacement
method just interprets the annotations.

It lacks Index support (for now) and its attribute naming logic should
evolve to only fall back to name parsing when an attribute-defining
annotation does not set the "attributeName" field.

Attribute-defining annotations are found on both fields and getter
methods.  On getter methods, the method must have "get", "is", or "has"
as a prefix, have zero arguments, and return a supported DynamoDb type.

The DynamoDBMarshaling annotation is not yet supported, but its not too
much labor to add support should we require it later.  Currently, we do
not use this yet.

LocalSecondary and Global index annotations are interpretted for the
sake of attribute definitions, but index definition itself is still
delegated to the subclasses.  This too can be resolved with a little
more effort.