Wrapping Transaction Library's transaction class to expose additional load() methods
The Transaction library excludes load(Object) and load(Object, Object),
the methods for fetch-by-HashKey and fetch-by-HashKey-and-RangeKey,
respectively.  There's no implementation reason why they couldn't be
included.  Rather than forking and modifying the transaction library
code base itself and folding that into our build, the shortcut of
subclassing it has been taken in the interest of time.