Proving how to fix Dynamo depenendencies
With both dynamo and JPA repositories in the application context path, a
JPA repository with the same name as a dynamo repository effectively
hides the dynamo repository and prevents it from being available for
Autowiring.

Fixed a test error that compalined that the dynamo
AchievementForUserRepository had no beans for autowire by disabling the
JPA repository of the same name.

Either we need to drop all duplicate repos or at least rename them.