Using pom inheritance to configure compiler and surefire plugins
One of maven's nifty features is the fact that child poms inherit from
parent poms.  I noticed that all the modules under mobsters-data-model
were repeating the same <build> configuration verbatim.

A change to that common configuration was made previously, but
mobsters-cache got skipped and so it had test cases that were not being
run.  Shared configurations are more readily changed correctly when
they are declared one time, not five.  Instead of just fixing the
mobsters-cache pom.xml, I moved the common build configuration for all
five modules to the mobsters-data-model parent pom.

It would not surprise me if something similar can be done with
mobsters-functional, I just have not attempted it there yet.