Phase One Refactoring for Dynamo Repository classes
-- Added distinction between the base repository for tables with a Hash
Key and those with a Hash+Range Key.
-- Refactored table DDL interface methods to a RepositorySetup interface
that is distinct from those used for Crud operations since these two
families of methods are targetted towards disjoint sets of consumers.
-- Corrected some inconsistencies between some Dynamo Entity annotations
and method names.

-- TEMPORARILY modified Entities using an integer to have two setter
methods, one of type int and the other of type Integer.  This is a
workarond concession to the use of a reflective approach to constructing
candidate key objects in the repositories that will disappear in the
next refactor phase.