A Director interface for GoF Builder/Director pattern conventionj
Controllers contribute a Director that encapsulates how to "build" a
description of the work they want a service to perform.  Services
contribute the concrete Builder that can interpret the construction of
this description appropriately.

Controllers call serviceMethod( directorImpl )
Services call director.apply( builderImpl )

Services then call any concrete-builder-defining build() methods that
are not part of the builder interface.  By enabling polymorphism in the
concrete builders, this design pattern lets the service reuse the
construction logic to build different kinds of output.  For example, a
service that embeds some other service might extend one of that
service's builder methods to add its own functionality and then provide
a concrete implementation that delegates that portion of the overall
interface to an implenentation acquired from the embedded service.