Added JavaDoc and replacing isEmpty() with isEmptyOrNull()
isNullOrEmpty() is the inverse of hasStorage().

While hasStorage() remains the preffered means of testing for positive
iterability, there are times when testing for negative iterability is
preferable (e.g. to branch into throwing an exception).

The static CollectionUtils.isEmpty() method was strictly inferior to the
Collection object method of the same name.  The static method required
indirection through a static utility library, but was equally prone to
causing NPE exceptions.  An such NPE exception would not be located at
the faulty guard check that allowed a null collection to "slip past",
making the root cause of such an error harder to find, hence static
CollectionUtils.isEmpty() was strictly inferior to non-static
Collection.isEmpty().

The method was renamed rather than just altering its behavior so that
CollectionUtils can remain viable as an XTend extension class by
avoiding name conflicts with a genuine object method.