Inverting MiniJobForUserService from Builder/Factory to Director/Builder design pattern style.
Instead of the Controller creating a Builder, configuring it, and
building its output to call the service with that output, this change
illustrates a use case where Controllers would create a Director that
encapsulates the code to configure a builder, call the service with that
encapsulation, and the service would construct the builder, pass it to
the director, and then consume the output of what the Builder
accumulated.

This is a more flexible design style for service-to-service delegation
and for constructing multiple outputs from the same logic.  For example,
if a controller's protobuf output merely echoes some part of its input
request when a service call succeeds, the controller may first pass its
Director to the service it wants to call, then use that same Director to
    process a Builder implementation of its own to construct the
Protobuf output.

FUTURE NOTES
============

When implementing a Director, the easiest thing to do is to process an
input protobuf during that Director's apply() method.  If the Director
is going to be reused multiple times, it may be favorible to use an
implementation of a DirectorBuilder.

For some Builder interface X, a DirectorBuilder is a mechanically
derived class that implements Director<X> and X.  It implements X by
accumulating enough data such that Director<X>.apply() can be
implemented by "replaying" the sequence of calls to its own Builder
methods to the interface of the Director.apply() method's argument
Builder.

If this is genuinely a need, the best way to get DirectorBuilder
implementations is by providing an XTend Active Annotation that will run
through the deterministic algorithm that yields the appropriate
implementation for any input interface X, where X is the single
interface implemented by an annotated class that is not Director<X>.