Configured a sample working EHCache entry.
Note: Not every field permitted by the XSD is really a supported <cache> property.  maxEntriesLocalHeap and maxEntriesInCache, for example, are permitted by the XSD, but are not defined in the EHCache configuration cache configuration class and so cause Exceptions when used.

maxElementsInMemory=0 creates an unrestricted cache, which seems the
fastest solution that works right now, given that configuration objects
are from a fixed size pool and cannot be created as a consequence of
user activity.