Using TimeUtils more consistently for unit test-friendly time handling
Only classes that are in a src/main directory and not located in a TODO
place are updated, but all such classes were swept over, exept for those
in the Dynamo and SQL projects.

The classes not touched are excluded because the TimeUtils class is not
in a sufficiently common project--the Dynamo and SQL modules simply
don't have it on their classpath to use yet.