Phase One DynamoDB Query Interface Design
This check in introduces the framework support for entity-specific Query
builders and demonstrates a manually-constructed user of that
framework on the QuestForUserRepository.

The framework's feature set is not fully implemented, but enough
features to demonstrate its use on the QuestForUser entity type are
completed and tested.  For code examples, see
E2EQuestForUserRepositoryQueryFilter.xtend.

The amount of incremental code required to use the framework with a
given entity type is small, but it is boilerplate and a pain to maintain
as an Entity changes over time.  The complexity was factored into a set
of framework classes to make it easier to generate this boilerplate at
build time, rather than hand implementing it as was done for this Proof
of Concept.

What's missing?
-- Support for Scan Filters (not clear whether or not we'll be using
Scan for anything yet.  It's a minor incremental add if we do)
-- Support for all DynamoDB types (only int and boolean are implemented
here.  Completing the required type support will not be very hard.)
-- Support for enabling "OR"-joined filter conditions.  A minor
interface addition to add a means of specifying "AND" or "OR" will be
required.  Until then only the default of "AND" is available.
-- DSL-driven code generation to keep Entity definitions and
FilterBuilder interfaces and implementations all consistent with one
another and spare the manual boilerplate, however small it may be.
-- Routing RangeKey filter conditions to the pre-filter RangeCondition
clause rather than the FilterCondition clause.  This will affect the
Repository implementation details, but should not affect the developer
interface to consume Query filters.
-- Index considerations.  Likely this will manifest as multiple Index-
specic Filtering interfaces to account for Projections and alternate
HashKeys, but we shall see...

NOTE: We are getting close to being able tosupport all Dynamo repo use cases without any hand written code beyond the entity+indexes DSL-formatted description.  It might be worth reconsidering the design constraint about following Spring-Data naming conventions as this will create a hand-written layer over top of the generated content to maintain.