Incremental maintenance to Conditions utility.
Added the ability to pass Logger LOG as an argument to use the same
message in both a thrown Exception and a logged error when a conditional
check fails and triggers a thrown Lvl6Exception.

Migrating to XTend next to facilitate lambdas as a source of
String creation.  I've encountered some multi-stage string construction
scenarios that defy simple Template + args deferal and don't want to
spend time constructing strings that won't get used in the common case
when a check passes.