Introduced "ExtensibleObject" abstraction.
An extensible object can have "Attachments" that are transient
state it carries at runtime but does not persist.

Implemented using Composition in order to be flexible with the
mode of selecting thread safety, although for now the implementation
does not leverage this flexibility.  An Atomic update strategy is
hard coded, but on late reflection, the ThreadPrivate might be fine
for Dynamo entities...

Deployed in only one object at present--StructureForUser

This will get reused in the History tracking example to follow.