Encourage use of the fluent query builder API to cut redundant code testing debt.
QuestForUserRepository recently recieved a condition-based query that is
readily adressed by using the fluent query builder API already present.

The test case that is the only consumer of this method would find it
easier to use if it were coded in XTend.  That refactor is not attempted
here, but a code comment is left behind to provide the one-liner that
would replace the block of code changed here.