DISABLING a test case that needs a way to create test data that fits the source code's MySQL Schema
This is a specific example of a general problem that is an obstacle to
writing configuration-based test cases right now.

The schema as known by the AWS branch does not match anything currently
in Dev or Staging.

A "World Fixture" utility module has been allocated, with the intent of
developing it into a "Test Utility" module that implements the "Shared
Fixture" or "Standard Fixture" test design patterns and is exposed to
test cases through one "Creation Method" and a suite of "Finder
Methods", but it is not sufficiently developped enough to serve those
roles yet.

For definitions of "Test Utility", "Shared Fixture", "Standard Fixture",
"Creation Methods", and "Finder Methods", see:

http://xunitpatterns.com/Test%20Utility%20Method.html
http://xunitpatterns.com/Shared%20Fixture.html
http://xunitpatterns.com/Standard%20Fixture.html
http://xunitpatterns.com/Creation%20Method.html
http://xunitpatterns.com/Test%20Utility%20Method.html#Finder%20Method