Test cases demonstrating how to use annotation-based validation of service call parameter syntax
The brief description goes as follows:

1)  Create a service that subclasses AbstractService

2)  Change service calls to return either an IRunnableAction or an
ICallableAction, depending on whether or not there will be result data
to return beyond a status code (which now goes to a
Lvl6MobstersException if it isn't RESULT_SUCCESS anyhow).  Whichever
interface you chose to implement, you must use AbstractAction as your
base type.

2a)  Pass all input arguments to a constructore on the new
AbstractAction subclass and store them in final fields of the subclass.

2b)  Next, add all @Autowired dependencies needed to service the service
call to the action's constructor method and store them in a similar
manner as the service call parameters.

2c)  Next, pass the service itself into the constructor.  This is the
only constructor argument you do NOT store a reference to in the
AbstractAction subclass.  Instead, pass it to AbstractAction's
superclass constructor (which will expect only the service's base type,
AbstractService).

2d)  Move the service implementation logic for the call that will
receieve validation logic inside the execute() method.  The execute
method needs no parameters because they are stored in the object thanks
to (2a) and (2b).

3)  Annotate the fields you added in (2a) to express any syntax rules
you want checked.  There are lots of built in annotations, and we can
implement custom ones both by writing code and by naming combinations of
existing validation annotations.

P.S. For an exmample of a good candidate for creating a named group of
constraints, look at the ID field of the "DummyAction" test object.  It
describes three rules -- not null, length = 32 characters, only
characters 0-9 and a-f permitted.  Collectively, these describe (or are
close to describing) a UUID.  UUIDs occur frequently enough that it
would be easier to just apply a @UUID annotation than it would be to
repeat all three of these constraints (or a corrected version thereof)
exactly the same way each time.

P.P.S. Email addresses (more sophisticated use of a regex),
configuration IDs (must always be positive integers), name fields (which
have a standard maximum length and are never empty), and description
fields (which have a standard maximum length that is larger than what
you see with names) are all good candidates for named constraint
combinations too!