The PaginatedQuestList's we've been returning to NOT support ListIterator
This causes problems when they end up the target of most anything
iterative, such as ImmutableList.copyOf(), for( X : List), and many
other common constructs.  This is a bit of a kludge, and it
should probably be implemented in query() itself because I don't think
I fixed every downstream consumer.

This unblocked me and its better, but I'll provide a more reliable fix
soon.