The PaginatedQuestList's we've been returning to NOT support ListIterator
This causes inappropriate UnsupportedOperationExceptions when they end
up the target of most anything iterative, such as
ImmutableList.copyOf(), for( X : List), and many other common
constructs.  This is a bit of a kludge, and it should probably be
implemented in query() itself because I don't think I could have
accounted for every potential downstream consumer.

This unblocked the code path I was struggling through at the time tho',
and its better, but I'll follow-up with a more appropriate fix soon.