Temporary work-around for ClosedSession exceptions
Lazy loading references must be traversed within a @Transactional
annotated method that keeps the Hibernate session open long enough
for them to not fail due to a lack of an open session.

The more significant problem is the necessity of having an open
Hibernate session when accessing configuration data.  For now, I
unblocked work by making all associations in a few key configuration
types load with a FetchType of EAGER, but this is just a band-aid.

The more long term solution will be to load an index required
configuration root objects on their first access through
ConfigExtensions and to manually trigger all lazy references
while building this index.