Marking business logic dependency as @Transient to avoid serialization
Also added a TODO note to favor equivalent implementations in
ConfigExtensions due to :

1)  Overhead of singleton references in
non-singleton artifacts.
2)  No way to provide access to the required dependency injection
without polluting ITask with implementation details.

(2) could be addressed by using spring-data's facility for adding
custom logic to a generated repository to have the repository
accept an injected singleton and relay it to every constructed
object on return, but it seems like a lot of work when most
consumers will have access to the configuration extensions
library anyhow.