Not all of the beans in this project are getting created and wired.
Added <component-scan> so Spring would locate and wire the Protobuf
to/fram Dynamo conversion utilities and websocket outbound byte array
service.

Marked validator as not lazy-initializing so it would b in the registry
for its autowire canidate tag to be effective.

Finally, set default-autowired-candidate="*" and
default-lazy-init="false" for the sake of component scan beans that
rely on those defaults ot be instantiated and injectable.