Maintenance to STOMP and Application Context XSD's
STOMP had some unused schemas included and ApplicationContext used
the wrong versions of context and beans.  Pruned the unused content
in the former case and up'd the versions to 4.0 in the latter case.

spring-stomp.xml is a somewhat wrong name choice.  This context
layer is really all about messaging, it just happens that exploratory
work with STOMP was the first piece added.  Renaming TBD when more of
its content is stable.