Websocket Handshake Handler
Making the default explicit, because this will need to get
overriden in order to access user credentials from the HTTP
context at the right time.  We must override a method here to
populate a spring-security Principal that will later identify
the user when broker, application, or other client needs to
route messages to just that user.