Message Converters.
This is an example Message ByteBuffer payload converter that produces
protobuf objects as output.  It will get replaced with a version
that Art wrote but I didn't have handy when I filled this in.

The implementation here does not require any fore-knowledge about
the available protobufs, nor does it need header data telling it
what class of ProtoBuf a given ByteBuffer contains.  Unfortuantely,
it also maintains a dynamically changing HashMap to retain what it
learns about each ProtoBuf class it sees for the first time in
order to be able to handle it faster should it be seen again.

Art's class is also capable of extracting all our known Protobufs
in a single class, but it uses up-front registration of those
classes to built the regisry immutably all botstrap and
uses an O(1) array lookup instead of O(log(n) hashmaps.

It also does not require references to Class objects, which is a
bonus Java best practices feature detail.