Prototype/WIP Websocket Handshake Handler
This will be used to place some Principal in the session context
so Spring will prepare itself to provide a facility to have a
message object serialized, encoded, encapsulated with headers,
and then sent back to its paired websocket client for delivery..

A real implementation should get the credential for a real user
from the HTTP context, not a random UUID.

Any method annotated with @MessageMapping to handle an input Message
may use @SendToUser to route its return message to a given use.

Any object whatsoever may have Spring wire it with a
org.springframework.messaging.simp.SimplMessageTemplate and call
convertAndSendToUser() in order to send a Message to a user
connected by WebSocket.

The name needed to target a websocket
user by either of these mechanisms is the name of the
Principal return by the HandshakeHandler method overriden here.
That name is provided through any Message that originated from
that user's WebSocket connection under the MessageHeader keyed
by SimpMessageHeaderAccessor.USER_HEADER.