Correcting the queue prefix list for relay routing
The prefixes must all begin and end with a slash.  In most cases, Spring
is being forgiving by appending a trailing "/" where one was expected
but not found, however I've already bumped into a bug with the user
prefix where Spring neither compensates for a missing trailing slash nor
flags it as an error, but instead yields broken functionality.

Rather than relying on a forgiving configuraiton interface, best to just
get things right explicitly in the first place.