Including a rudimentary exception handling layer.
Still missing is the construction of a Reply object that triggered the
result, but that is most likely going to be implemented in the game
state webapp, not here.  This will simply relay the message as it is
already doing.

The exception to that generalization would be if an exception blocks
message delivery to the game state server.  We'll revisit that in due
time.  It's good to get something routing to a logger for now.