Relocating ConfigExtensions to the domain model project
When its config object functionality is separated to a
ConfigurationRegistry, it will still need visibility on that split
artifact because it will still have logic that must retrieve
configuration objects.

The ConfigurationRegistry will live in the domain project so the CQAttr
artifacts it requires need not be populated within the mobsters-sql
project itself.  Even though they must be created in the
com.lvl6.mobsters.info java package to function, they are not relevant
to relational access and therefore do not belong in the same java
module.