Removing @Component from four Dynamo repos to explicitly configure their default read/write units
spring-dynamo.xml-related changes were previously submitted in another
check in and ammont to explicitly instnatiating Dynamo repository beans
in order to inject each with a specific ProvisioningSpec.

It was not possible to do this with @Qualifier and <qualifier> because
all repositories inherit their provisioning spec from a common base
class that can only bear one annotation value.