Resolving errors in the replacement configuration metadata cache.
This shaved 100ms off earlier testing that was reporting 280ms (but was
not fully processing the command unfortunately).

"Initialization on Holder Class Load Event" is a design pattern for
managing the initialization of singletons with a post-constructor init()
step that has to be accomplished while supporting multi-threaded
consumers.

What makes it necessary here is the fact that when initialization state
is not mutated post-initialization, consumers do not need
synchronization to see the initialized state.

What makes it a little challenging to work with here is the fact that
the ClassLoader's thread is being used to do initialization work and
that thread does not obey breakpoints when a debugger is being used,
making it potentially difficult to observe some of the code effects
unless you run the initialization logic outside of its Holder.