Merging the MobstersMeta configuration model wrappers with the MobsterPlayer document model.
The instinct to place these in separate packages was sound and meant to
partition the root atifact into meaningful subunits.  XCore sometimes
handles the cross-file dependencies correctly when deriving .ecore
artifacts, but it somewhat frequently does not.  It's proving to be more
trouble than its worth, so this check-in re-merges the two packages to
one.

The XCore compiler does not claim to support multi-package usage yet,
but perhaps it will catch up on a later release.  It's a project closely
related to the more mature XTend and XText projects, but is still
somewhat young, so this need be a lesson to only use features advertised
as complete.  :)