Dividing the new domain model project into two
The XCore framework is related to XText and XTend, but Maven support for
its code generator is still on the roadmap.

For the time being, its necessary to deal with these the same way that
we deal with Protobuf--by checkining in the generated source.  The
.xcore model file is still the authoritative source--when it changes,
the generated output is checked in accordingly.  There is purportedly
support for hand modification to the generated code in some places, but
there should not be any reason to ever attempt that!