Bi-directional protobuf conversion
The Mobsters encoding algorithm is NOT symmetric.  Header encoding is
BIG_ENDIAN for incoming requests but LITTLE_ENDIAN for outgoing replies.
There is a better way to account for that, but this approach was the
fastest to get up and going.

Also realized that since we're currently using the conversion algorithm
that requires an array of builder prototypes, that array also needed to
be built for reply objects.  The initial pass here is not complete, but
its enough to keep move forward.