Correcting one facet of the AMQP/Redis non-functioning issue
THe primary reason the demo does not work correctly with AMQP/Redis is a
missing data flow
path between the strict-order-output channel and the to-output-tracking
channel that begins the
last stage of processing the chrolonologically ordered event stream.

The JMS variant sends this channel's output to a message queue that
every worker node consumes from.  A s
single worker threead picks up and consumes tasks without having to
check worry about how to
to, allowing workers to pick up the "next" task and proceed on the
assumption that overall ordering of all tasks is
chronologically valid.

This approach mimics that, but uses an AMQP queue in place of a JMS
message queue.  The missing stage enqueues its input, and there was
already a pre-existing stage attempting to pick up messages from a queue
that nothing was
writing to yet.  It needed a little minor correction, but was already
close enough to this current form to indicate what was intended.